# docs/studio/openapi/studio-api.yaml
openapi: 3.0.3
info:
  title: OpenWorkflow Studio Workspace API
  version: 1.0.0
  description: >-
    Versioned workspace API for listing, validating, and safely persisting canonical
    OpenWorkflow and OpenAPI catalog source documents.
servers:
  - url: /
    description: Same-origin Quarkus application
...
security:
  - bearerAuth: []
paths:
...
  /api/studio/v1/sync:
    get:
      tags: [Synchronization]
      operationId: getSyncStatus
      summary: Read canonical-to-generated artifact status
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Workspace synchronization status
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Studio-API-Version: { $ref: '#/components/headers/ApiVersion' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncStatus' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Synchronization]
      operationId: requestSync
      summary: Retry or coalesce generated-artifact synchronization
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                operationId: { type: string, maxLength: 128 }
                documentIds:
                  type: array
                  maxItems: 200
                  items: { type: string }
      responses:
        '202':
          description: Synchronization queued
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Studio-API-Version: { $ref: '#/components/headers/ApiVersion' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncStatus' }
        default: { $ref: '#/components/responses/Problem' }
  /api/studio/v1/trash/{trashId}/restore:
    parameters:
      - name: trashId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 128 }
    post:
      tags: [Recovery]
      operationId: restoreTrashEntry
      summary: Restore a deleted document to its original path
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IfNoneMatchRequired'
      responses:
        '201':
          description: Restored document
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Studio-API-Version: { $ref: '#/components/headers/ApiVersion' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '409': { $ref: '#/components/responses/DuplicatePath' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        default: { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, maxLength: 128, pattern: '^[\x20-\x7E]+$' }
...
    IfNoneMatchRequired:
      name: If-None-Match
      in: header
      required: true
      schema: { type: string, maxLength: 200, example: '*' }
...
  headers:
    ETag:
      description: Quoted SHA-256 hash of exact source bytes.
      schema: { type: string }
    RequestId:
      description: Caller-provided or server-generated request identifier.
      schema: { type: string }
    ApiVersion:
      description: Selected Studio API contract version.
      schema: { type: string, example: '1' }
  responses:
    Problem:
      description: Normalized problem response
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        X-Studio-API-Version: { $ref: '#/components/headers/ApiVersion' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    DuplicatePath:
      description: Requested path already exists
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    PreconditionRequired:
      description: Required concurrency precondition was omitted
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    DocumentKind:
      type: string
      enum: [workflow, catalog]
    DocumentFormat:
      type: string
      enum: [yaml, json]
    CompatibilityMode:
      type: string
      enum: [editable, partial, source-readonly]
    GenerationState:
      type: string
      enum: [pending, in_sync, out_of_sync, blocked, disabled]
    GenerationStatus:
      type: object
      required: [state]
      properties:
        state: { $ref: '#/components/schemas/GenerationState' }
        operationId: { type: string }
        lastAttempt: { type: string, format: date-time }
        message: { type: string }
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/GeneratedArtifact' }
    GeneratedArtifact:
      type: object
      required: [path, state]
      properties:
        path: { type: string }
        state: { $ref: '#/components/schemas/GenerationState' }
        sourceEtag: { type: string }
        generatedEtag: { type: string }
        lastAttempt: { type: string, format: date-time }
        message: { type: string }
    DocumentSummary:
      type: object
      required: [id, kind, path, format, sizeBytes, etag, compatibility, generation]
      properties:
        id: { type: string }
        kind: { $ref: '#/components/schemas/DocumentKind' }
        path: { type: string }
        displayName: { type: string }
        format: { $ref: '#/components/schemas/DocumentFormat' }
        sizeBytes: { type: integer, minimum: 0 }
        etag: { type: string }
        revisionNumber: { type: integer, minimum: 1 }
        modifiedAt: { type: string, format: date-time }
        compatibility: { $ref: '#/components/schemas/CompatibilityMode' }
        specVersion: { type: string }
        openapi: { type: string }
        documentVersion: { type: string }
        name: { type: string }
        workflowId: { type: string }
        stateTypes:
          type: array
          items: { type: string }
        catalogAliases:
          type: array
          items: { type: string }
        functionReferences:
          type: array
          items: { type: string }
        catalogReferences:
          type: array
          items: { type: string }
        subflowReferences:
          type: array
          items: { type: string }
        reusableSubflow: { type: boolean }
        parseStatus:
          type: string
          enum: [parsed, error]
        validationState:
          type: string
          enum: [valid, parse-error, unsupported]
        generation: { $ref: '#/components/schemas/GenerationStatus' }
        diagnostics:
          type: array
          items: { $ref: '#/components/schemas/Diagnostic' }
    Document:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          required: [content]
          properties:
            content: { type: string, maxLength: 2097152 }
            metadata:
              $ref: '#/components/schemas/DocumentMetadata'
            sourceTree:
              description: Parsed source tree used for safe generic details rendering.
              type: object
              additionalProperties: true
...
    NamedItem:
      type: object
      required: [name]
      properties:
        name: { type: string }
        detail: { type: string }
    DocumentMetadata:
      type: object
      properties:
        workflowId: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: string }
        specVersion: { type: string }
        openapi: { type: string }
        start: { type: string }
        timeouts: { type: object, additionalProperties: true }
        constants: { type: object, additionalProperties: true }
        annotations: { type: object, additionalProperties: true }
        extensions: { type: array, items: { type: object, additionalProperties: true } }
        functions: { type: array, items: { $ref: '#/components/schemas/NamedItem' } }
        events: { type: array, items: { $ref: '#/components/schemas/NamedItem' } }
        errors: { type: array, items: { $ref: '#/components/schemas/NamedItem' } }
        stateCounts: { type: object, additionalProperties: { type: integer } }
        terminalStates: { type: array, items: { type: string } }
        subflowReferences: { type: array, items: { type: string } }
...
    Diagnostic:
      type: object
      required: [id, ruleId, phase, severity, message, provenance, documentationUrl, suppressible]
      properties:
        id: { type: string }
        ruleId: { type: string }
        phase:
          type: string
          enum: [parse, schema, semantic, compatibility, runtime]
        severity:
          type: string
          enum: [error, warning, info, hint]
        message: { type: string }
        explanation: { type: string }
        suggestedResolution: { type: string }
        primaryRange: { $ref: '#/components/schemas/SourceRange' }
        relatedRanges:
          type: array
          items:
            type: object
            required: [range]
            properties:
              message: { type: string }
              range: { $ref: '#/components/schemas/SourceRange' }
        fieldPath: { type: string }
        nodeId: { type: string }
        documentationUrl: { type: string, format: uri }
        suppressible: { type: boolean }
        provenance:
          type: string
          enum: [local-parser, local-schema, local-analyzer, sonataflow]
    SourceRange:
      type: object
      required: [start, end, encoding]
      properties:
        start: { $ref: '#/components/schemas/Position' }
        end: { $ref: '#/components/schemas/Position' }
        encoding:
          type: string
          enum: [utf-16-code-units]
    Position:
      type: object
      required: [offset, line, column]
      properties:
        offset: { type: integer, minimum: 0 }
        line: { type: integer, minimum: 1 }
        column: { type: integer, minimum: 1 }
...
    SyncStatus:
      type: object
      required: [state, policy, artifacts]
      properties:
        state: { $ref: '#/components/schemas/GenerationState' }
        policy: { type: string, enum: [on-save] }
        operationId: { type: string }
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/GeneratedArtifact' }
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    Conflict:
      type: object
      required: [expectedEtag, actualEtag, merge]
      properties:
        expectedEtag: { type: string }
        actualEtag: { type: string }
        currentDocument: { $ref: '#/components/schemas/Document' }
        merge: { type: string, enum: [three-way-client] }
    Problem:
      type: object
      required: [type, title, status, code, detail, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        conflict: { $ref: '#/components/schemas/Conflict' }